--- src/app/missions/rocketsSlice.js
@@ -1,26 +1,0 @@
-import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
-
-const initialState = {
-  rockets: [],
-};
-
-const url = 'https://api.spacexdata.com/v4/rockets';
-
-export const fetchRockets = createAsyncThunk('rockets/fetchRockets', async () => {
-  const res = await fetch(url);
-  const data = await res.json();
-  return data;
-});
-
-export const rocketsSlice = createSlice({
-  name: 'rockets',
-  initialState,
-  reducers: {},
-  extraReducers: (builder) => {
-    builder.addCase(fetchRockets.fulfilled, (state, action) => {
-      state.rockets = action.payload;
-    });
-  },
-});
-
-export default rocketsSlice.reducer;
