+++ src/app/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 import missionsSlice from './missions/missionsSlice';
-import rocketsSlice from './missions/rocketsSlice';
+import rocketsSlice from './rockets/rocketsSlice';
 
 const store = configureStore({
   reducer: {
